Conflicts:
	ExpenseTracker/app/src/main/res/layout/layout_dialog_edit_text.xml
	ExpenseTracker/app/src/main/res/values/strings.xml

+++ ExpenseTracker/app/src/main/java/com/pedrocarrillo/expensetracker/custom/CrossfadePageTransformer.java
@@ -60,7 +60,7 @@ public class CrossfadePageTransformer implements ViewPager.PageTransformer {
             }
 
             if(ivCreditCard != null){
-                ivCreditCard.setTranslationX(-(float)(pageWidth * 0.5f * position));
+                ivCreditCard.setTranslationX(-(pageWidth * 0.5f * position));
             }
 
             if(ivMoneyPig != null){
